Dashboard 3.3.0 Catalyst
These classes are fragments specifically created for the Catalyst
Rom itself. The Init.d class is relatively new and may need
tweaked as we make changes to the kernel and init.d folder inside
the rom. please review what xml files run this class in the rom.
The KexecMenu activity is empty at the moment and was created for
future options for the JBX kernel. Kernel developers are free to
write to this file as well as the kexec.xml that is tied to this
class.